+++ packages/xchain-cosmos/src/types/client-types.ts
@@ -1,8 +1,4 @@
 import { Network } from '@xchainjs/xchain-client'
-import { Asset, Chain } from '@xchainjs/xchain-util'
-
-export const AssetAtom: Asset = { chain: Chain.Cosmos, symbol: 'ATOM', ticker: 'ATOM', synth: false }
-export const AssetMuon: Asset = { chain: Chain.Cosmos, symbol: 'MUON', ticker: 'MUON', synth: false }
 
 export type ClientUrls = Record<Network, string>